[frontend] re-organize context
